+++ plugin/src/main/java/com/jacobibanez/plugin/android/godotplaygameservices/snapshots/SnapshotsProxy.kt
@@ -1,0 +1,33 @@
+package com.jacobibanez.plugin.android.godotplaygameservices.snapshots
+
+import android.util.Log
+import androidx.core.app.ActivityCompat
+import com.google.android.gms.games.PlayGames
+import com.google.android.gms.games.SnapshotsClient
+import org.godotengine.godot.Godot
+
+/** @suppress */
+class SnapshotsProxy(
+    private val godot: Godot,
+    private val snapshotsClient: SnapshotsClient = PlayGames.getSnapshotsClient(godot.getActivity()!!)
+) {
+    private val tag = SnapshotsProxy::class.java.simpleName
+
+    private val showSavedGamesRequestCode = 9010
+
+    fun showSavedGames(
+        title: String,
+        allowAddButton: Boolean,
+        allowDelete: Boolean,
+        maxSnapshots: Int
+    ) {
+        Log.d(tag, "Showing save games")
+        snapshotsClient.getSelectSnapshotIntent(title, allowAddButton, allowDelete, maxSnapshots)
+            .addOnSuccessListener { intent ->
+                ActivityCompat.startActivityForResult(
+                    godot.getActivity()!!, intent,
+                    showSavedGamesRequestCode, null
+                )
+            }
+    }
+}